# swagger.yaml
openapi: 3.0.3
info:
  title: Bell Project API
  description: QR-based restaurant ordering system API
  version: 1.0.0
  contact:
    name: Bell Project Team
    email: support@bellproject.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.bellproject.com/dev
    description: Development server
  - url: https://api.bellproject.com/prod
    description: Production server
...
paths:
...
  /order:
    post:
      tags:
        - order
      summary: Create new order
      description: Create a new order with menu items and customer information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostOrderRequest'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostOrderResponse'
        '400':
          description: Bad request - Invalid order data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Menu not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '410':
          description: Menu item not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Unprocessable entity - Invalid menu item or price mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /order/{orderId}:
    get:
      tags:
        - order
      summary: Get order details
      description: Retrieve details for a specific order
      parameters:
        - name: orderId
          in: path
          required: true
          description: Unique order identifier
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
        - name: restaurantId
          in: query
          required: true
          description: Restaurant identifier for the order
          schema:
            type: string
            example: "rest_001"
      responses:
        '200':
          description: Order retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetOrderResponse'
        '400':
          description: Bad request - Missing required parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - order
      summary: Cancel order
      description: Cancel an existing order with potential refund calculation
      parameters:
        - name: orderId
          in: path
          required: true
          description: Unique order identifier
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
        - name: restaurantId
          in: query
          required: true
          description: Restaurant identifier for the order
          schema:
            type: string
            example: "rest_001"
      responses:
        '200':
          description: Order cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteOrderResponse'
        '400':
          description: Bad request - Cannot cancel completed order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    OrderItem:
      type: object
      required:
        - menuItemId
        - quantity
        - price
        - selectedOptions
      properties:
        menuItemId:
          type: string
          example: "item_001"
        quantity:
          type: integer
          minimum: 1
          example: 2
        price:
          type: number
          format: float
          minimum: 0
          example: 25.98
        selectedOptions:
          type: array
          items:
            $ref: '#/components/schemas/SelectedOption'
    SelectedOption:
      type: object
      required:
        - optionId
        - choiceId
        - priceModifier
      properties:
        optionId:
          type: string
          example: "size"
        choiceId:
          type: string
          example: "large"
        priceModifier:
          type: number
          format: float
          example: 2.00
    CustomerInfo:
      type: object
      required:
        - name
        - phone
      properties:
        name:
          type: string
          example: "John Doe"
        phone:
          type: string
          example: "010-1234-5678"
        email:
          type: string
          format: email
          example: "john@example.com"
...
    PostOrderRequest:
      type: object
      required:
        - restaurantId
        - items
        - customerInfo
      properties:
        restaurantId:
          type: string
          example: "rest_001"
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
          minItems: 1
        customerInfo:
          $ref: '#/components/schemas/CustomerInfo'
    PostOrderResponse:
      type: object
      properties:
        orderId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        status:
          type: string
          enum: [CREATED]
          example: "CREATED"
        totalAmount:
          type: number
          format: float
          example: 25.98
        paymentUrl:
          type: string
          format: uri
          example: "https://payment.example.com/pay?orderId=550e8400-e29b-41d4-a716-446655440000"
        createdAt:
          type: string
          format: date-time
          example: "2023-08-15T12:30:45.123Z"
    GetOrderResponse:
      type: object
      properties:
        orderId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        restaurantId:
          type: string
          example: "rest_001"
        status:
          type: string
          enum: [CREATED, PAID, COOKING, READY, COMPLETED, CANCELLED]
          example: "PAID"
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        totalAmount:
          type: number
          format: float
          example: 25.98
        createdAt:
          type: string
          format: date-time
          example: "2023-08-15T12:30:45.123Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-08-15T12:35:45.123Z"
        paymentInfo:
          type: object
          properties:
            method:
              type: string
              enum: [naverpay, kakaopay]
              example: "naverpay"
            transactionId:
              type: string
              example: "txn_123456789"
            paidAt:
              type: string
              format: date-time
              example: "2023-08-15T12:35:45.123Z"
    DeleteOrderResponse:
      type: object
      properties:
        orderId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        status:
          type: string
          enum: [CANCELLED]
          example: "CANCELLED"
        refundAmount:
          type: number
          format: float
          example: 25.98
        refundMethod:
          type: string
          example: "original"
...
    ErrorResponse:
      type: object
      required:
        - type
        - title
        - status
        - detail
      properties:
        type:
          type: string
          format: uri
          example: "https://httpstatuses.com/400"
        title:
          type: string
          example: "Bad Request"
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: "The request is missing required parameters"
        instance:
          type: string
          format: uri
          example: "/menu"